Extract and clarify standalone probe flow
The fact the Queue Proxy's main.go actually has a standalone exec probe mode,
and what it does (and why :)), is pretty confusing to new readers.

This makes it a bit clearer by adding some clarifying comments and pulling the
standalone-exec-probe stuff out to its own file.

Also removes unused failingHealthCheck var.
